requirements.txt
uuid
numpy
PyPDF2
chromadb-client
langchain
accelerate
ipykernel
torch
flashrank
python-dotenv
transformers
sentence-transformers